% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_color.R
\name{prim_dark_blue}
\alias{prim_dark_blue}
\alias{prim_med_blue}
\alias{prim_pink}
\alias{prim_teal}
\title{Individual OBI Color}
\usage{
prim_dark_blue(show_color = F)

prim_med_blue(show_color = F)

prim_pink(show_color = F)

prim_teal(show_color = F)
}
\arguments{
\item{show_color}{if display the color}
}
\description{
All OBI colors
}
