% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_palettes.R
\name{primary_colors_3}
\alias{primary_colors_3}
\alias{secondary_colors_6}
\alias{tertiary_colors_7}
\alias{all_colors}
\title{OBI color palettes}
\usage{
primary_colors_3(show_color = F)

secondary_colors_6(show_color = F)

tertiary_colors_7(show_color = F)

all_colors(show_color = F)
}
\arguments{
\item{show_color}{}
}
\description{
colors from OBI style guide
}
\examples{
library(tidyverse)
ggplot(data = iris,
       aes(x = Sepal.Length, y = Petal.Length, color = Species)) +
  geom_point() +
  # use OBI colors
  scale_color_manual(values = primary_colors_3())
}
