% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_color.R
\name{prim_blue}
\alias{prim_blue}
\alias{prim_purple}
\alias{prim_green}
\title{Individual OBI Color}
\usage{
prim_blue(show_color = F)

prim_purple(show_color = F)

prim_green(show_color = F)
}
\arguments{
\item{show_color}{if display the color}
}
\description{
All OBI colors
}
